fix dtypes issue with compositing speed up
Creating the composited image can be sped up by only performing
calculations where there is a label to overlay. ndimage.find_objects
must have an input array with an int dtype, which is not guaranteed
in npz files (or necessary for other functionality, like actions).
Cast input array to int for find_objects call to get this to work
for float dtype label files without changing the content in the npz