recomposite pixel-editing image if label cmap changes
This causes all labels to shift color, and only updating a small
portion of the image causes a visual bug, especially with the prism
label cmap. Adds a flag to CalibanWindow, toggled in ZStackReview
handle_draw and handle_threshold, that determines when composite
needs to be redrawn for all labels (when max label in file changes).